mod: system to handle textbox has been modified in the header in the gui

+++ include/gui/InfoGame.hpp
@@ -202,7 +202,7 @@ namespace Gui
     struct InfoSlider {
         BoxSlider sliderBar;
         BoxSlider sliderHandle;
-        double sliderValue;
+        int sliderValue;
     };
 }
